Some code cleanups to EFormidling types
* Added interface and default to get receivers
* Moved types to separate namespace
* Removed unused NullImplementation
